requirements.txt
orange3-text
whoosh
amcatclient>=3.4.9
progressmonitor>=0.5
